refactor(context): mark/relocate single-context files in network/ and sync/ (#10088)
Follow-up to the execution-context boundary work. Investigation found no
genuine renderer/node fork in these folders that warrants a new runtime
adapter — the network fork already has its adapter (node uses curlRequest/
script-executor, renderer uses cancellation/concurrency), and git is
main-only behind IPC with renderer type-only imports. So the fix is the
existing suffix convention + relocation, not new adapters.

network/ (suffix the 4 single-context outliers so they opt into the
existing *.renderer/*.node ESLint rules):
- cancellation.ts -> cancellation.renderer.ts (uses window.main)
- concurrency.ts  -> concurrency.renderer.ts  (uses window.main)
- send-request.ts -> send-request.node.ts     (node:fs, inso CLI only)
- grpc/write-proto-file.ts -> .node.ts         (node:fs/os/path, main only)

sync/ (move the lone renderer file out of the node-heavy folder):
- sync/vcs/insomnia-sync.ts -> ui/utils/insomnia-sync.ts (window.main.sync
  + imports ui/ components), matching the ui/utils/ precedent. Test moved
  alongside.

No runtime behaviour change; imports updated at all call sites.
