+++ packages/insomnia/src/runtimes/network/network-adapter.renderer.ts
@@ -7,8 +7,8 @@ import type { RenderedRequest } from '~/templating/types';
 import type { RequestContext } from '../../../../insomnia-scripting-environment/src/objects';
 import type { CurlRequestOptions, ResponsePatch } from '../../main/network/libcurl-promise';
 import { applyDefaultHeaders } from '../../network/apply-default-headers';
-import { cancellableCurlRequest } from '../../network/cancellation';
-import { runScriptConcurrently } from '../../network/concurrency';
+import { cancellableCurlRequest } from '../../network/cancellation.renderer';
+import { runScriptConcurrently } from '../../network/concurrency.renderer';
 
 export const getTimelinePath = (responseId: string): Promise<string> => window.main.timeline.getPath(responseId);
